Turn core/std lib into a crate
