+++ library/src/tests/canon.rs
@@ -3,7 +3,7 @@
 
 #![allow(clippy::needless_raw_string_hashes)]
 
-use crate::{test_expression, test_expression_with_lib};
+use super::{test_expression, test_expression_with_lib};
 use indoc::indoc;
 use qsc::interpret::Value;
