+++ library/src/tests/diagnostics.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::test_expression;
+use super::test_expression;
 use qsc::interpret::Value;
 
 #[test]
