+++ library/src/tests/measurement.rs
@@ -3,7 +3,7 @@
 
 #![allow(clippy::needless_raw_string_hashes)]
 
-use crate::test_expression;
+use super::test_expression;
 use indoc::indoc;
 use qsc::interpret::Value;
