+++ library/src/tests/resources/compare.qs
