+++ library/src/tests/resources/inc_by_le.qs
