+++ library/src/tests/resources/qft_le.qs
