--- library/tests/src/tests.rs
@@ -1,103 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#![allow(clippy::needless_raw_string_hashes)]
-
-use crate::{test_expression, test_expression_with_lib_and_profile};
-use indoc::indoc;
-use qsc::{interpret::Value, target::Profile};
-
-//
-// Core namespace
-//
-#[test]
-fn check_repeated() {
-    test_expression("Repeated(Zero, 0)", &Value::Array(vec![].into()));
-    test_expression(
-        "Repeated(One, 1)",
-        &Value::Array(vec![Value::RESULT_ONE].into()),
-    );
-    test_expression(
-        "Repeated(1, 2)",
-        &Value::Array(vec![Value::Int(1), Value::Int(1)].into()),
-    );
-    test_expression(
-        "Repeated(true, 3)",
-        &Value::Array(vec![Value::Bool(true), Value::Bool(true), Value::Bool(true)].into()),
-    );
-}
-
-#[test]
-fn check_exp_with_cnot() {
-    // This decomposition only holds if the magnitude of the angle used in Exp is correct and if the
-    // sign convention between Rx, Rz, and Exp is consistent.
-    test_expression(
-        indoc! {r#"{
-            open Microsoft.Quantum.Diagnostics;
-            open Microsoft.Quantum.Math;
-
-            use (aux, control, target) = (Qubit(), Qubit(), Qubit());
-            within {
-                H(aux);
-                CNOT(aux, control);
-                CNOT(aux, target);
-            }
-            apply {
-                let theta  = PI() / 4.0;
-                Rx(-2.0 * theta, target);
-                Rz(-2.0 * theta, control);
-                Adjoint Exp([PauliZ, PauliX], theta, [control, target]);
-
-                Adjoint CNOT(control, target);
-            }
-
-            CheckAllZero([aux, control, target])
-        }"#},
-        &Value::Bool(true),
-    );
-}
-
-#[test]
-fn check_exp_with_swap() {
-    // This decomposition only holds if the magnitude of the angle used in Exp is correct.
-    test_expression(
-        indoc! {r#"{
-            open Microsoft.Quantum.Diagnostics;
-            open Microsoft.Quantum.Math;
-
-            use (aux, qs) = (Qubit(), Qubit[2]);
-            within {
-                H(aux);
-                CNOT(aux, qs[0]);
-                CNOT(aux, qs[1]);
-            }
-            apply {
-                let theta  = PI() / 4.0;
-                Exp([PauliX, PauliX], theta, qs);
-                Exp([PauliY, PauliY], theta, qs);
-                Exp([PauliZ, PauliZ], theta, qs);
-
-                Adjoint SWAP(qs[0], qs[1]);
-            }
-
-            CheckAllZero([aux] + qs)
-        }"#},
-        &Value::Bool(true),
-    );
-}
-
-#[test]
-fn check_base_profile_measure_resets_aux_qubits() {
-    test_expression_with_lib_and_profile(
-        indoc! {"{
-            use q = Qubit();
-            X(q);
-            let result = M(q);
-            Reset(q);
-            result
-        }"},
-        "",
-        Profile::Base,
-        &Value::RESULT_ONE,
-    );
-}
